refactor + fix getPromptInfoFromId
